splitted log exporter into LogHelper class
